frameless_window: Hidden frame on MacOS and enable drag
Tutorial is clearly MacOS focused and chooses the easy solution for
hiding a frame on MacOS that doesn't hide the frame elsewhere.

The sass updates enable dragging to return on MacOS but dragging isn't
missing on Windows so it is hard to view the differences. Selection of
<h1> tags is to be disabled, but that doesn't occur. Took the step of
comparing changes to demo repo, but with no differences the text is
still selectable. It isn't selectable in the demo repo so maybe a step
is missed in the tutorial?